polkadot_node_primitives::approval::time
Trait Clock

pub trait Clock {
    // Required methods
    fn tick_now(&self) -> Tick;
    fn wait(
        &self,
        tick: Tick,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'static>>;
}

A clock which allows querying of the current tick as well as waiting for a tick to be reached.

Required Methods

fn tick_now(&self) -> Tick

Yields the current tick.

fn wait(&self, tick: Tick) -> Pin<Box<dyn Future<Output = ()> + Send + 'static>>

Yields a future which concludes when the given tick is reached.