Crate polkadot_node_primitives

Primitive types used on the node-side.

Unlike the polkadot-primitives crate, these primitives are only used on the node-side, not shared between the node and the runtime. This crate builds on top of the primitives defined there.

Re-exports

pub use disputes::CandidateVotes;

pub use disputes::InvalidDisputeVote;

pub use disputes::SignedDisputeStatement;

pub use disputes::ValidDisputeVote;

Modules

approval

Types relevant for approval.

disputes

Disputes related types.

Macros

new_session_window_size

Create a new checked SessionWindowSize

Structs

AvailableData

This is the data we keep available for each candidate included in the relay chain.

BabeEpoch

BABE epoch information

BabeEpochConfiguration

Configuration data used by the BABE consensus engine that may change with epochs.

BlockData

Parachain block data.

Collation

The output of a collator.

CollationGenerationConfig

Configuration for the collation generator

CollationResult

Result of the CollatorFn invocation.

CollationSecondedSignal

Signal that is being returned when a collation was seconded by a validator.

DisputeMessage

A dispute initiating/participating message that is guaranteed to have been built from signed statements.

ErasureChunk

A chunk of erasure-encoded block data.

PoV

A Proof-of-Validity

Proof

This is a convenience type to allow the Erasure chunk proof to Decode into a nested BoundedVec

SessionWindowSize

Type of a session window size.

UncheckedDisputeMessage

A DisputeMessage where signatures of statements have not yet been checked.

Enums

BabeAllowedSlots

Types of allowed slots.

DisputeMessageCheckError

Things that can go wrong when constructing a DisputeMessage.

InvalidCandidate

Candidate invalidity details

MaybeCompressedPoV

A type that represents a maybe compressed PoV.

MerkleProofError

Possible errors when converting from Vec<Vec<u8>> into Proof.

Statement

A statement, where the candidate receipt is included in the Seconded variant.

ValidationResult

Result of the validation of the candidate.

Constants

APPROVAL_EXECUTION_TIMEOUT

The amount of time to spend on execution during approval or disputes.

BACKING_EXECUTION_TIMEOUT

The amount of time to spend on execution during backing.

DISPUTE_WINDOW

It would be nice to draw this from the chain state, but we have no tools for it right now. On Polkadot this is 1 day, and on Kusama it’s 6 hours.

MAX_FINALITY_LAG

Linked to MAX_FINALITY_LAG in relay chain selection, MAX_HEADS_LOOK_BACK in approval-voting and MAX_BATCH_SCRAPE_ANCESTORS in dispute-coordinator

POV_BOMB_LIMIT

The bomb limit for decompressing PoV blobs.

VALIDATION_CODE_BOMB_LIMIT

The bomb limit for decompressing code blobs.

Functions

maybe_compress_pov

Compress a PoV, unless it exceeds the POV_BOMB_LIMIT.

Type Definitions

BlockWeight

The cumulative weight of a block in a fork-choice rule.

CollatorFn

Collation function.

SignedFullStatement

A statement, the corresponding signature, and the index of the sender.

UncheckedSignedFullStatement

Variant of SignedFullStatement where the signature has not yet been verified.