polkadot_node_primitives::approval

Module time

Time utilities for approval voting. Time utilities for approval voting subsystems.

Structs

DelayedApprovalTimer

A list of delayed futures that gets triggered when the waiting time has expired and it is time to sign the candidate. We have a timer per relay-chain block.

SystemClock

A clock which uses the actual underlying system clock.

Constants

TICK_DURATION_MILLIS

The duration of a single tick in milliseconds.

TICK_TOO_FAR_IN_FUTURE

How far in the future a tick can be accepted.

Traits

Clock

A clock which allows querying of the current tick as well as waiting for a tick to be reached.

ClockExt

Extension methods for clocks.

Functions

slot_number_to_tick

assumes slot_duration_millis evenly divided by tick duration.

tick_to_slot_number

Converts a tick to the slot number.

tranche_to_tick

Converts a tranche from a slot to the tick number.

Type Aliases

Tick

A base unit of time, starting from the Unix epoch, split into half-second intervals.