Module polkadot_node_primitives::approval::v2

A list of primitives introduced by v2.

Structs

AssignmentCertV2

A certification of assignment.

BitIndex

A bit index in Bitfield.

Bitfield

A read-only bitvec wrapper

IndirectAssignmentCertV2

An assignment criterion which refers to the candidate under which the assignment is relevant by block hash.

Slot

Unit type wrapper that represents a slot.

VrfOutput

VRF output type suitable for schnorrkel operations.

VrfProof

VRF proof type suitable for schnorrkel operations.

VrfSignature

VRF signature data

VrfTranscript

Transcript ready to be used for VRF related operations.

Enums

AssignmentCertKindV2

Certificate is changed compared to AssignmentCertKind:

AssignmentConversionError

Errors that can occur when trying to convert to/from assignment v1/v2

BitfieldError

Errors that can occur when creating and manipulating bitfields.

Constants

ASSIGNED_CORE_CONTEXT

A static context associated with producing randomness for v2 multi-core assignments.

CORE_RANDOMNESS_CONTEXT

A static context associated with producing randomness for a core.

RELAY_VRF_MODULO_CONTEXT

A static context used for all relay-vrf-modulo VRFs for v2 multi-core assignments.

Traits

AsBitIndex

Helper trait to convert primitives to BitIndex.

Type Definitions

CandidateBitfield

A read-only, non-zero bitfield. Each 1 bit identifies a candidate by the bitfield bit index.

CoreBitfield

A bitfield of core assignments.

Randomness

Randomness type required by BABE operations.