Module polkadot_node_primitives::disputes

Disputes related types.

Structs

CandidateVotes
Tracked votes on candidates, for the purposes of dispute resolution.

DisputeMessage
A dispute initiating/participating message that is guaranteed to have been built from signed statements.

InvalidDisputeVote
Any invalid vote (currently only explicit).

SignedDisputeStatement
A checked dispute statement from an associated validator.

UncheckedDisputeMessage
A DisputeMessage where signatures of statements have not yet been checked.

ValidDisputeVote
Any valid vote (backing, approval, explicit).

Enums

DisputeMessageCheckError
Things that can go wrong when constructing a DisputeMessage.